updates vision app and adds a complementary java subsystem
this commit updates the vision python application to make it
easier to understand and use. It also adds a vision subsystem
that needs to be added to the robot code to allow the python
application to work. An example of its implementation can be
found in Team 1100's FRCChargedUp repository.